+++ src/main/java/com/project/ewdj/repository/BookRepository.java
@@ -7,8 +7,10 @@ import org.springframework.stereotype.Repository;
 import com.project.ewdj.entity.Book;
 
 @Repository
-public interface BookRepository extends CrudRepository<Book, Integer> {
+public interface BookRepository extends CrudRepository<Book, Long> {
 
     List<Book> findByBookName(String bookName);
     // Book findByISBN(String isbn_nummer);
+
+    // public Book assignAuthorToBook(Long book_id, Long author_id);
 }
